+++ bothub_cli/main.py
@@ -94,13 +94,15 @@ def add_option_to_dict(d, option_name, option):
 @click.option('--api-key')
 @click.option('--app-id')
 @click.option('--app-secret')
-def add_channel(channel, api_key, app_id, app_secret):
+@click.option('--page-access-token')
+def add_channel(channel, api_key, app_id, app_secret, page_access_token):
     '''Add a new channel to current project'''
     try:
         credentials = {}
         add_option_to_dict(credentials, 'api_key', api_key)
         add_option_to_dict(credentials, 'app_id', app_id)
         add_option_to_dict(credentials, 'app_secret', app_secret)
+        add_option_to_dict(credentials, 'page_access_token', page_access_token)
         lib.add_channel(channel, credentials)
     except exc.CliException as ex:
         click.secho('{}: {}'.format(ex.__class__.__name__, ex), fg='red')